Fixes an issue when more than one external ID field matches on create
To reproduce, log into salesforce and change an existing Contact's
(not yet in Web 2.0) `Web Front End Email` to match the Contact's
email address.  Then change that Contact's `Web Front End ID` to `1`
or a value known to duplicate an internal ID in Web 2.0.  Then,
create an account in Web 2.0 with that email address.  An exception
will be thrown due to not being able to store the ID field (too long).