Use Rails.cache to save extra queries to Salesforce
RecordType needs to be queried to resolve the ID.  Since we don't
store this value, it must be queried every time.  This change uses
Rails.cache to cache the value for 1 day.  This should reduce the
number of calls to Salesforce.  This change also includes using
the exception wrapper so that exceptions for this are reported by
ExceptionNotifier emails.