[CDS-30] Add action to selected field
The current code to export licenses use action to count the login
activity from the user to be exported but it was not in the selected
field of the query. I add it and the problem was fix it.

The next steps could be optimization of the query or send it to am
asynchronous task.