[CDS-794] Add styles and classes following BEM methodology
This is a proposal of how we should handle the components of the CMS.

The first point is start using [BEM](http://getbem.com/introduction/)
methodology to write the sass code and classes of each component.

The second point is create a sass file for each component on this
folder: `app/assets/stylesheets/components/cms/`.